+++ kwsCheckSemicolonSpace.cxx
@@ -27,7 +27,7 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
   delete [] val;
 
   bool hasError = false;
-  long int posSemicolon = m_BufferNoComment.find(";",0);
+  long int posSemicolon = static_cast<long int>(m_BufferNoComment.find(";",0));
   while(posSemicolon != -1)
     {
     // We try to find the word before that
@@ -73,11 +73,11 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
         }
       i--;
       }
-    posSemicolon = m_BufferNoComment.find(";",posSemicolon+1);
+    posSemicolon = static_cast<long int>(m_BufferNoComment.find(";",posSemicolon+1));
     }
 
   // We also check if we have two or more semicolons following each other
-  posSemicolon = m_BufferNoComment.find(";",0);
+  posSemicolon = static_cast<long int>(m_BufferNoComment.find(";",0));
   while(posSemicolon != -1)
     {
     // We look forward
@@ -115,12 +115,12 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
         }
       i++;
       }
-    posSemicolon = m_BufferNoComment.find(";",i+1);
+    posSemicolon = static_cast<long int>(m_BufferNoComment.find(";",i+1));
     }
 
 
   // We also check if we have unecessary semicolons
-  posSemicolon = m_BufferNoComment.find(";",0);
+  posSemicolon = static_cast<long int>(m_BufferNoComment.find(";",0));
   while(posSemicolon != -1)
     {
     // check if we have a } before
@@ -138,9 +138,9 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
       {
       bool error = true;
       // We check that this is not a class
-      long int openingChar = this->FindOpeningChar('}','{',pos,true);
+      long int openingChar = static_cast<long int>(this->FindOpeningChar('}','{',pos,true));
 
-      long int classPos = m_BufferNoComment.find("class");
+      long int classPos = static_cast<long int>(m_BufferNoComment.find("class"));
       while(classPos != -1)
         {
         if(classPos != -1 && openingChar!= -1)
@@ -158,11 +158,11 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
             error = false;
             }
           }
-        classPos = m_BufferNoComment.find("class",classPos+1);
+        classPos = static_cast<long int>(m_BufferNoComment.find("class",classPos+1));
         }
 
       // We check that this is not a enum
-      long int enumPos = m_BufferNoComment.find("enum");
+      long int enumPos = static_cast<long int>(m_BufferNoComment.find("enum"));
       while(enumPos != -1)
         {
         if(enumPos != -1 && openingChar!= -1)
@@ -180,7 +180,7 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
             error = false;
             }
           }
-        enumPos = m_BufferNoComment.find("enum",enumPos+1);
+        enumPos = static_cast<long int>(m_BufferNoComment.find("enum",enumPos+1));
         }
 
       std::string word = this->FindPreviousWord(openingChar);
@@ -190,7 +190,7 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
         }
 
       // We check that this is not a struct
-      long int structPos = m_BufferNoComment.find("struct");
+      long int structPos = static_cast<long int>(m_BufferNoComment.find("struct"));
       while(structPos != -1)
         {
         if(structPos != -1 && openingChar!= -1)
@@ -208,7 +208,7 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
             error = false;
             }
           }
-        structPos = m_BufferNoComment.find("struct",structPos+1);
+        structPos = static_cast<long int>(m_BufferNoComment.find("struct",structPos+1));
         }
 
       word = this->FindPreviousWord(openingChar);
@@ -218,7 +218,7 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
         }
 
       // We check that this is not a union
-      long int unionPos = m_BufferNoComment.find("union");
+      long int unionPos = static_cast<long int>(m_BufferNoComment.find("union"));
       while(unionPos != -1)
         {
         if(unionPos != -1 && openingChar!= -1)
@@ -236,7 +236,7 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
             error = false;
             }
           }
-        unionPos = m_BufferNoComment.find("union",unionPos+1);
+        unionPos = static_cast<long int>(m_BufferNoComment.find("union",unionPos+1));
         }
 
       word = this->FindPreviousWord(openingChar);
@@ -268,7 +268,7 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
         {
         if(m_BufferNoComment[i]=='}')
           {
-          long int localopeningChar = this->FindOpeningChar('}','{',i,true);
+          long int localopeningChar = static_cast<long int>(this->FindOpeningChar('}','{',i,true));
           long int j = localopeningChar-1;
           while(j>0 && (m_BufferNoComment[j]==' '
             || m_BufferNoComment[j]=='\n'
@@ -317,7 +317,7 @@ bool Parser::CheckSemicolonSpace(unsigned long max)
         hasError = true;
         }
       }
-    posSemicolon = m_BufferNoComment.find(";",posSemicolon+1);
+    posSemicolon = static_cast<long int>(m_BufferNoComment.find(";",posSemicolon+1));
     }
 
   return !hasError;
